# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v2/auth/public/time/": {
            "get": {
                "description": "Auth Server Current Unix Timestamp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Get Server Current Unix Timestamp",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                        }
                    },
                    "400": {
                        "description": "Failed",
                        "schema": {
                            "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v2/auth/public/version/": {
            "get": {
                "description": "Auth Version",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Get Auth Version",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                        }
                    },
                    "400": {
                        "description": "Failed",
                        "schema": {
                            "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v2/auth/resource/documents/": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a Document",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Create a Document",
                "parameters": [
                    {
                        "type": "string",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "docCategory",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Create a file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Document response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/user-service_api_dto.DocumentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v2/auth/resource/documents/{id}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get a Document",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Get a Document",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/user-service_api_dto.DocumentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
        "/v2/auth/resource/locations/": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get a Location",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Locations"
                ],
                "summary": "Get a Location",
                "responses": {
                    "200": {
                        "description": "Location response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/user-service_api_dto.LocationResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update a Location",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Locations"
                ],
                "summary": "Update a Location",
                "parameters": [
                    {
                        "description": "Update a Location",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user-service_api_dto.UpdateLocationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Location response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/user-service_api_dto.LocationResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a Location",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Locations"
                ],
                "summary": "Create a Location",
                "parameters": [
                    {
                        "description": "Create a Location",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user-service_api_dto.CreateLocationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Location response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/user-service_api_dto.LocationResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "user-service_api_dto.CreateLocationRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 255
                },
                "city": {
                    "type": "string",
                    "maxLength": 50
                },
                "country": {
                    "type": "string",
                    "maxLength": 50
                },
                "metadata": {
                    "type": "string",
                    "maxLength": 255
                },
                "postalCode": {
                    "type": "string",
                    "maxLength": 10
                },
                "staticPhoneNumber": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
...
        "user-service_api_dto.DocumentResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "directory": {
                    "type": "string"
                },
                "docCategory": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mimeType": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "user-service_api_dto.LocationResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "metadata": {
                    "type": "string"
                },
                "postalCode": {
                    "type": "string"
                },
                "staticPhoneNumber": {
                    "type": "string"
                },
                "verification": {
                    "type": "string"
                }
            }
        },
...
        "user-service_api_dto.UpdateLocationRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 255
                },
                "city": {
                    "type": "string",
                    "maxLength": 50
                },
                "country": {
                    "type": "string",
                    "maxLength": 50
                },
                "metadata": {
                    "type": "string",
                    "maxLength": 255
                },
                "postalCode": {
                    "type": "string",
                    "maxLength": 10
                },
                "staticPhoneNumber": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
...
        "user-service_api_helper.BaseHttpResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "result": {},
                "resultCode": {
                    "$ref": "#/definitions/user-service_api_helper.ResultCode"
                },
                "success": {
                    "type": "boolean"
                },
                "validationErrors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user-service_api_validation.ValidationError"
                    }
                }
            }
        },
        "user-service_api_helper.ResultCode": {
            "type": "integer",
            "enum": [
                0,
                40001,
                40101,
                40301,
                40401,
                42901,
                42902,
                50001,
                50002
            ],
            "x-enum-varnames": [
                "Success",
                "ValidationError",
                "AuthError",
                "ForbiddenError",
                "NotFoundError",
                "LimiterError",
                "OtpLimiterError",
                "CustomRecovery",
                "InternalError"
            ]
        },
        "user-service_api_validation.ValidationError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "property": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AuthBearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}